+++ test/components/OnlineSearchForm.test.tsx
@@ -51,7 +51,6 @@ describe('OnlineSearchForm', () => {
 		expect(screen.getByRole('button', { name: 'Search' })).not.toBeNull();
 		expect(screen.getByRole('button', { name: 'Clear' })).not.toBeNull();
 		expect(screen.getByRole('button', { name: 'Close' })).not.toBeNull();
-		expect(screen.getByRole('button', { name: 'Save Search' })).not.toBeNull();
 		expect(screen.getByRole('radio', { name: 'Asc' })).toBeChecked();
 		expect(screen.getByRole('radio', { name: 'Desc' })).not.toBeChecked();
 	});